api/rethinkdb-flask-restful-api/requirements.txt
Flask-Migrate==2.7.0
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Jinja2==2.11.3
Werkzeug==1.0.1
flasgger==0.9.5
itsdangerous==1.1.0
pytz==2021.1
requests==2.25.1
setproctitle==1.1.10
six==1.12.0
